+++ lib/output-utils.sh
@@ -46,7 +46,7 @@ function inf() {
 }
 
 function info-debug() {
-  [[ -z ${DEBUG} ]] && return
+  [[ -z ${BASHMATIC_DEBUG} ]] && return
   printf -- "${LibOutput__LeftPrefix}${bakpur}[ debug ] $*  ${clr}\n"
 }
 
@@ -123,19 +123,20 @@ function columnize() {
 }
 
 function dbg-on() {
-  export DEBUG=$(time.now.db)
+  export BASHMATIC_DEBUG=$(time.now.db)
 }
 
 function dbg-off() {
-  unset DEBUG
+  unset BASHMATIC_DEBUG
+  [[ -f ${BASHMATIC_HOME}/.envrc.no-debug ]] && source ${BASHMATIC_HOME}/.envrc.no-debug
 }
 
 # @description Checks if we have debug mode enabled
 function is-dbg() {
-  [[ -n $DEBUG ]]
+  [[ -n ${BASHMATIC_DEBUG} ]]
 }
 
-# @description Local debugging helper, activate it with DEBUG=1
+# @description Local debugging helper, activate it with `export BASHMATIC_DEBUG=1`
 function dbg() {
   is-dbg && printf "     ${txtgrn}[DEBUG | ${txtylw}$(time.now.with-ms)${txtgrn}]  ${txtblu}$(txt-info)$*\n" >&2
   return 0