+++ test/helpers/test-source.sh
@@ -1,4 +1,6 @@
 #!/usr/bin/env bash
 
-source "${BASHMATIC_INIT}"
-bashmatic.validate-sourced-in
+source "${BASHMATIC_INIT}" refresh
+source "${BASHMATIC_HOME}/lib/subshell.sh" 
+
+bashmatic.subshell.validate-sourced-in